feat: Adiciona a função CalculateAge ao Utilitaries
