+++ system/backend/src/utilitaries/utilitaries.service.ts
@@ -24,4 +24,20 @@ export class UtilitariesService {
     }
     return true;
   }
+
+  calculateAge(birthday: string | Date): number {
+    const birthDate = new Date(birthday);
+    const today = new Date();
+
+    let age = today.getFullYear() - birthDate.getFullYear();
+    const monthDiff = today.getMonth() - birthDate.getMonth();
+    const dayDiff = today.getDate() - birthDate.getDate();
+
+    // Adjust if the birthday hasn't occurred yet this year
+    if (monthDiff < 0 || (monthDiff === 0 && dayDiff < 0)) {
+      age--;
+    }
+
+    return age;
+  }
 }
